vibe_check_python_backend/pyproject.toml
[project]
name = "vibe-check-python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.2.0",
    "pydantic-settings>=2.14.1",
    "uvicorn>=0.46.0",
]
